Revert "Enable BuiltInNotifications by default on iOS" https://bugs.webkit.org/show_bug.cgi?id=278062 rdar://problem/133795669
Reviewed by Sihui Liu.

Enabling built-in notifications by default on iOS (282148@main) did not work as expected because of
how it interacts with DeprecatedGlobalSettings. The issue is that in UIProcess, nothing actually
calls `DeprecatedGlobalSettings::setBuiltInNotifications(true)` even if you flip the default in the
YAML file. Meanwhile, in WebProcess, `WebPage::updatePreferencesGenerated` does call
`DeprecatedGlobalSettings::setBuiltInNotifications(true)`. So the value of
`DeprecatedGlobalSettings::builtInNotificationsEnabled()` is different in WebContent and UIProcess,
which leads to very weird behavior (and actually some message checks start firing).

We need to revert this and come up with a better way of changing the default.

* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebKit/NetworkProcess/NetworkProcess.cpp:
(WebKit::NetworkProcess::processPushMessage):

Canonical link: https://commits.webkit.org/282218@main