Remove `svgDocument` interface from `HTMLFrameElement` to align with other browsers
https://bugs.webkit.org/show_bug.cgi?id=278069
rdar://133800070

Reviewed by NOBODY (OOPS!).

This patch aligns WebKit with Web Specification [1]"

[1] https://html.spec.whatwg.org/multipage/embedded-content-other.html#svg-0

As per web specification, `getSVGDocument` should only be exposed on
'iframe', 'embed' and 'object'. This patch removes it from 'frame' element.

Additionally, we made our implementation to return null in 259905@main but
never synced our IDL interface for 'object' and 'embed', so this does and match
it with 'iframe'.

* Source/WebCore/html/HTMLEmbedElement.idl:
* Source/WebCore/html/HTMLFrameElement.idl:
* Source/WebCore/html/HTMLObjectElement.idl:
* LayoutTests/svg/custom/frame-getSVGDocument-expected.txt:
* LayoutTests/svg/custom/frame-getSVGDocument.html: