[JSC] Should use canBeInternal to check the node intended for fusing in B3LowerToAir https://bugs.webkit.org/show_bug.cgi?id=257132 rdar://109666425
Reviewed by NOBODY (OOPS!).

Given the pattern

@3 = BitXor(@1, -1)
@4 = BitAnd(@0, @3)

for BIC, we want to convert it to

@4 = BIC(@0, @1)

However, if @3 is also used elsewhere besides @4 then we would emit

@3 = BitXor(@1, -1)
@4 = BIC(@0, @1)

In that case, node BitXor would be recomputed again. To avoid that
scenario, we should use canBeInternal(@3) to confirm whether the
fusion is beneficial. This patch fixes the this issue in B3LowerToAir.

* Source/JavaScriptCore/b3/B3LowerToAir.cpp: