[GTK][WPE] WebProcessPool::sandboxEnabled() not always consistent https://bugs.webkit.org/show_bug.cgi?id=278158
Reviewed by Adrian Perez de Castro.

We decide whether to enabled sandbox or not when launching the process,
so that sandbox can be enabled in WebProcessPool but we still disable it
because we are running under valgrind or it's been disabled by an env
var. We should decide whether to enable the sandbox or nor in
WebProcessPool and then when launching the process just check the
launch options.

* Source/WebKit/UIProcess/Launcher/glib/ProcessLauncherGLib.cpp:
(WebKit::ProcessLauncher::launchProcess):
(WebKit::isSandboxEnabled): Deleted.
* Source/WebKit/UIProcess/WebProcessPool.h:
* Source/WebKit/UIProcess/glib/WebProcessPoolGLib.cpp:
(WebKit::WebProcessPool::platformInitialize):
(WebKit::WebProcessPool::setSandboxEnabled):

Canonical link: https://commits.webkit.org/282281@main