[WPE][GTK] Crash in WebCore::TextDecorationPainter::paintBackgroundDecorations when compiled with Clang with LTO enabled https://bugs.webkit.org/show_bug.cgi?id=274780
Reviewed by Michael Catanzaro.

Clang seem to have some issue when inlining assignment and move
operators in LTO builds, generating code that tries to perform OOB
access to Vector data. Replacing an assignmenmt with a Vector::swap(),
which is semantically equivalent in this case (the moved-from object
is not used again in the function) workarounds the compiler issue.

* Source/WebCore/rendering/TextDecorationPainter.cpp:
(WebCore::translateIntersectionPointsToSkipInkBoundaries):

Canonical link: https://commits.webkit.org/282306@main