Avoid repeat calling for Node::renderer() in FragmentDirectiveRangeFinder https://bugs.webkit.org/show_bug.cgi?id=278232
Reviewed by NOBODY (OOPS!).

A comment in Node.h suggests to avoid repeated calls to Node::renderer(),
as it includes branching. A few places in FrameDirectiveRangeFinder.cpp
have repeat calls to ::renderer() that can be improved.

Additionally, use a CheckedPtr to store the RenderObject, as calls
to RenderObject::style() can be non-trivial depending on the implementation.

* Source/WebCore/dom/FragmentDirectiveRangeFinder.cpp:
(WebCore::FragmentDirectiveRangeFinder::nearestBlockAncestor):
(WebCore::FragmentDirectiveRangeFinder::isVisibleTextNode):
(WebCore::FragmentDirectiveRangeFinder::rangeOfStringInRange):