Fix broken ENABLE_MALLOC_HEAP_BREAKDOWN feature. Bug URL: https://bugs.webkit.org/show_bug.cgi?id=278260 Radar Link: rdar://134090283
Reviewed by NOBODY (OOPS!).

Add a missing DECLARE_ALLOCATOR_WITH_HEAP_IDENTIFIER macro to FrameRateAligner.h.
When WTF_MAKE_FAST_ALLOCATED_WITH_HEAP_IDENTIFIER is used, AcceleratedTimeLine.h relies on a class defined in AnimationMalloc.h, but could not find it when compiling. Put in a missing include statement for AnimationMalloc.h and change the target membership for AnimationMalloc.h to be in WebCore. This fixed the problem.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/animation/AcceleratedTimeline.h:
* Source/WebCore/animation/FrameRateAligner.h: