AX: Support the AXRangesForSearchPredicate API that VoiceOver uses to retrieve misspellings. https://bugs.webkit.org/show_bug.cgi?id=278254 <rdar://problem/134083674>
Reviewed by Tyler Wilcock.

Follow up to https://github.com/WebKit/WebKit/pull/31136.

This functionality was originally implemented as an extension of the AXUIElementsForSearchPredicate API, but that didn't work for VoiceOver since it changes the type of the return value which is unexpected for the middleware system framework. Therfore, a new AXRangesForSearchPredicate API was added to expose the functionality on the Mac.

The corresponding AccessibilityUIElement::textMarkerRangeForSearchPredicate test method was updated.

* Source/WebCore/accessibility/AXSearchManager.cpp:
(WebCore::AXSearchManager::findMatchingRange):
* Source/WebCore/accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper accessibilityAttributeValue:forParameter:]):
* Tools/WebKitTestRunner/InjectedBundle/mac/AccessibilityUIElementMac.mm:
(WTR::AccessibilityUIElement::textMarkerRangeForSearchPredicate):

Canonical link: https://commits.webkit.org/282401@main