Unreviewed, suppress wrong threading assertion for InlineCacheHandler https://bugs.webkit.org/show_bug.cgi?id=278435 rdar://134290611
It gets destroyed in the GC end phase and it gets stale assertion about threading checks. But that assertion is wrong since
GC end phase is ensuring that we are not running mutator, so we can safely and correctly derefing this InlineCacheHandler inside it.
This is fixing this wrong assertion.

* Source/JavaScriptCore/bytecode/InlineCacheCompiler.cpp:
(JSC::m_watchpoint):
* Source/JavaScriptCore/bytecode/InlineCacheCompiler.h:

Canonical link: https://commits.webkit.org/282538@main