[Unified PDF] Embedded PDFs are blank in snapshots https://bugs.webkit.org/show_bug.cgi?id=278420 rdar://132092364
Reviewed by Simon Fraser and Abrar Rahman Protyasha.

258399@main made us opt explicitly into local snapshot buffers for PDFPlugin
in the main frame, but not for subframes. We need local buffers in both cases,
since UnifiedPDFPlugin depends on having a platform context to paint into.

Check for PluginView (which is always PDFPlugin) instead of checking that the
frame itself is plugin-backed.

* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::snapshotAtSize):
(WebKit::snapshotImageOptions): Deleted.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/UnifiedPDFTests.mm:
(TestWebKitAPI::configurationForWebViewTestingUnifiedPDF):
(TestWebKitAPI::TEST(UnifiedPDF, SnapshotsPaintPageContent)):
Add a test that confirms that the PDF content is painted.

Canonical link: https://commits.webkit.org/282551@main