Disable use of JSC signal handlers by default on Apple embedded platforms. https://bugs.webkit.org/show_bug.cgi?id=278401 rdar://134053581
Reviewed by NOBODY (OOPS!).

This is because mach exception handlers are only allowed by the WebContent process
sandbox.  So, we'll change the default behavior to disallow its use, and explicitly
allow options that depend on it only for the WebContent process.  These options
are: usePollingTraps, and useWasmFaultSignalHandler.

Additionally, when enabling these options, we'll also have to re-enable useWasmFastMemory.
This is because useWasmFastMemory could have been disabled due to useWasmFaultSignalHandler
defaulting to false.  After emabling useWasmFaultSignalHandler, we'll also need to
enable useWasmFastMemory to give it a chance to stay enabled.

We'll also need to enable these options in the jsc shell, testapi, and DumpRenderTree
because their initialization does not use XPCServiceInitializer.

Also, XPCServiceInitializer is used by the WebContent process as well as the GPU and
Network processes.  To only enable the mach exception handler dependent options for
WebContent process and not the others, we need to add a canAllowSignalHandlers template
argument to XPCServiceInitializer.  Only WebContent process sets this argument to true.

* Source/JavaScriptCore/API/tests/testapi.cpp:
(configureJSCForTesting):
* Source/JavaScriptCore/jsc.cpp:
(CommandLine::parseArguments):
* Source/JavaScriptCore/runtime/InitializeThreading.cpp:
(JSC::initialize):
* Source/JavaScriptCore/runtime/OptionsList.h:
* Source/JavaScriptCore/wasm/WasmFaultSignalHandler.cpp:
(JSC::Wasm::prepareSignalingMemory):
* Source/WebKit/Shared/EntryPointUtilities/Cocoa/XPCService/XPCServiceEntryPoint.h:
(WebKit::XPCServiceInitializer):
* Source/WebKit/WebProcess/EntryPoint/Cocoa/XPCService/WebContentServiceEntryPoint.mm:
(WEBCONTENT_SERVICE_INITIALIZER):
* Tools/DumpRenderTree/mac/DumpRenderTree.mm:
(dumpRenderTree):